+++ native/CHANGELOG.md
@@ -1,5 +1,13 @@
 # wy-react-native
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - wy-helper@1.1.1
+  - wy-react-helper@1.0.1
+
 ## 2.0.0
 
 ### Patch Changes